Add support for shell installation on Windows
This adds support for installing the Windows version of Deno from the
shell script which should work as long as the user has a basic UNIX like
environment (MSYS, WSL, MingW, et cetera).

It works for GitHub runners as-well which means there's less friction in
setting up a matrix which installs Deno.